+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/immutableArraysModule/ImmutableArrayGenerator.kt
@@ -291,6 +291,15 @@ private fun generateImmutableArrayFile(baseType: BaseType): FileSpec {
                 returns = Int::class.asTypeName(),
                 forceFunctionBody = true,
             )
+            "indexOfLast"(
+                typeSpecBuilder = this,
+                baseType = baseType,
+                parameters = {
+                    "predicate"(type = lambda<Boolean> { "element"(type = baseType.type) })
+                },
+                returns = Int::class.asTypeName(),
+                forceFunctionBody = true,
+            )
             addPartition(baseType)
             addSortedBy(baseType)
             addSortedByDescending(baseType)