+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -283,6 +283,13 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * See [Array.indexOfLast]
+     */
+    public inline fun indexOfLast(predicate: (element: T) -> Boolean): Int {
+        return values.indexOfLast(predicate)
+    }
+
+    /**
      * Creates a pair of immutable arrays, where the first contains elements for which predicate
      * yielded true, and the second contains the other elements.
      */