+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -288,6 +288,13 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
     }
 
     /**
+     * See [BooleanArray.indexOfLast]
+     */
+    public inline fun indexOfLast(predicate: (element: Boolean) -> Boolean): Int {
+        return values.indexOfLast(predicate)
+    }
+
+    /**
      * Creates a pair of immutable arrays, where the first contains elements for which predicate
      * yielded true, and the second contains the other elements.
      */